requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
psycopg2-binary==2.9.9
jinja2==3.1.2
python-dotenv==1.0.0
google-generativeai==0.3.2
httpx==0.25.2
aiofiles==23.2.1
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
schedule==1.2.0